+++ src/adaptors/index.ts
@@ -26,6 +26,7 @@ export { default as getPoapData } from './poap';
 export { default as checkPoH } from './poh';
 export { default as getPolygonData } from './polygon';
 export { default as getProjectGalaxyData } from './projectgalaxy';
+export { default as getPopData } from './proofofpersonhood';
 export { default as getRabbitholeData } from './rabbithole';
 export { default as getRaribleData } from './rarible';
 export { default as getRss3Data } from './rss3';